advtask

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -9,19 +9,18 @@
 void print_python_list_info(PyObject *p)
 {
 	int i;
-	Py_ssize_t size;
-	Py_ssize_t alloc;
+	int size;
+	int alloc;
 	PyObject *obj;
 
 	size = Py_SIZE(p);
 	alloc = ((PyListObject *)p)->allocated;
-	printf("[*] Size of the Python List = %d\n", (int)size);
-	printf("[*] Allocated = %d\n", (int)alloc);
+	printf("[*] Size of the Python List = %d\n", size);
+	printf("[*] Allocated = %d\n", alloc);
 
 	for (i = 0; i < size; i++)
 	{
 		printf("Element %d: ", i);
-
 		obj = PyList_GetItem(p, i);
 		printf("%s\n", Py_TYPE(obj)->tp_name);
 	}
